+++ Frontend/my-app/package.json
@@ -23,11 +23,14 @@
     "antd": "^5.25.1",
     "axios": "^1.9.0",
     "bootstrap": "^5.3.6",
+    "chart.js": "^4.4.9",
+    "dayjs": "^1.11.13",
     "gsap": "^3.13.0",
     "lucide-react": "^0.510.0",
     "normalize.css": "^8.0.1",
     "react": "^18.3.1",
     "react-calendar": "^5.1.0",
+    "react-chartjs-2": "^5.3.0",
     "react-dom": "^18.3.1",
     "react-redux": "^9.2.0",
     "react-router-dom": "^7.6.0",
