--- Frontend/my-app/src/components/OrderCardList/OrderCardList.scss
@@ -1,33 +1,0 @@
-.cards-grid {
-  display: grid;
-  grid-template-columns: repeat(1, 1fr);
-  position: relative;
-  padding-top: 5px;
-
-  @media (min-width: 640px) {
-    grid-template-columns: repeat(2, 1fr);
-  }
-
-  @media (min-width: 1024px) {
-    grid-template-columns: repeat(3, 1fr);
-  }
-
-  gap: 1.5rem;
-  max-width: 1120px;
-  margin: 0 auto;
-}
-
-.scroll-container {
-  max-height: calc(100vh - 140px);
-  overflow-y: auto;
-  padding: 20px 16px 80px;
-  box-sizing: border-box;
-
-  /* רק הסתרת סרגל הגלילה בלי לשנות עיצוב אחר */
-  scrollbar-width: none; 
-  -ms-overflow-style: none;  
-
-  &::-webkit-scrollbar {
-    display: none; 
-  }
-}
