+++ Frontend/my-app/src/components/commonComponents/MovieCard/MovieCard.scss
