+++ Frontend/my-app/src/components/commonComponents/SearchBar/SearchBar.tsx
