+++ Frontend/my-app/src/components/commonComponents/UserCard/UserCard.scss
