+++ Frontend/my-app/src/components/commonComponents/UserCardList/UserCardList.scss
