+++ Frontend/my-app/src/components/userComponents/Camera/Camera.tsx
