+++ Frontend/my-app/src/components/userComponents/HomePage/HomePage.scss
